airport_ns/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.2
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
drf-jwt==1.19.2
gunicorn==20.1.0
mysqlclient==2.1.0
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
